move edgecases
